# api_gateway/api/docs/swagger.yaml
definitions:
...
  user_service.CreateUsersRequest:
    properties:
      branch_id:
        type: integer
      firstname:
        type: string
      lastname:
        type: string
      login:
        type: string
      password:
        type: string
      phone:
        type: string
    type: object
...
  user_service.ListUsersResponse:
    properties:
      count:
        type: integer
      users:
        items:
          $ref: '#/definitions/user_service.Users'
        type: array
    type: object
  user_service.Response:
    properties:
      message:
        type: string
    type: object
...
  user_service.UpdateUsersRequest:
    properties:
      branch_id:
        type: integer
      firstname:
        type: string
      id:
        type: integer
      lastname:
        type: string
      login:
        type: string
      password:
        type: string
      phone:
        type: string
    type: object
  user_service.Users:
    properties:
      active:
        type: boolean
      branch_id:
        type: integer
      created_at:
        type: string
      firstname:
        type: string
      id:
        type: integer
      lastname:
        type: string
      login:
        type: string
      password:
        type: string
      phone:
        type: string
      updated_at:
        type: string
    type: object
  v1.Response:
    properties:
      data: {}
      description:
        type: string
      status:
        type: integer
    type: object
info:
  contact: {}
paths:
...
  /v1/user/create:
    post:
      consumes:
      - application/json
      description: Create a new user with the provided details
      parameters:
      - description: data of the user
        in: body
        name: user
        required: true
        schema:
          $ref: '#/definitions/user_service.CreateUsersRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/user_service.Response'
        "400":
          description: Bad Request
          schema:
            allOf:
            - $ref: '#/definitions/v1.Response'
            - properties:
                data:
                  type: string
              type: object
        "404":
          description: Not Found
          schema:
            allOf:
            - $ref: '#/definitions/v1.Response'
            - properties:
                data:
                  type: string
              type: object
        "500":
          description: Internal Server Error
          schema:
            allOf:
            - $ref: '#/definitions/v1.Response'
            - properties:
                data:
                  type: string
              type: object
      summary: Create a new user
      tags:
      - user
  /v1/user/delete/{id}:
    delete:
      consumes:
      - application/json
      description: delete a user by its unique identifier
      parameters:
      - description: Catgeory ID to retrieve
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/user_service.Response'
        "400":
          description: Bad Request
          schema:
            allOf:
            - $ref: '#/definitions/v1.Response'
            - properties:
                data:
                  type: string
              type: object
        "404":
          description: Not Found
          schema:
            allOf:
            - $ref: '#/definitions/v1.Response'
            - properties:
                data:
                  type: string
              type: object
        "500":
          description: Internal Server Error
          schema:
            allOf:
            - $ref: '#/definitions/v1.Response'
            - properties:
                data:
                  type: string
              type: object
      summary: Delete a Catgory
      tags:
      - user
  /v1/user/get/{id}:
    get:
      consumes:
      - application/json
      description: Retrieve a user by its unique identifier
      parameters:
      - description: User ID to retrieve
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/user_service.Users'
        "400":
          description: Bad Request
          schema:
            allOf:
            - $ref: '#/definitions/v1.Response'
            - properties:
                data:
                  type: string
              type: object
        "404":
          description: Not Found
          schema:
            allOf:
            - $ref: '#/definitions/v1.Response'
            - properties:
                data:
                  type: string
              type: object
        "500":
          description: Internal Server Error
          schema:
            allOf:
            - $ref: '#/definitions/v1.Response'
            - properties:
                data:
                  type: string
              type: object
      summary: Get a user by ID
      tags:
      - user
  /v1/user/list:
    get:
      consumes:
      - application/json
      description: get user
      parameters:
      - default: 10
        description: limit for response
        in: query
        name: limit
        type: integer
      - default: 1
        description: page for response
        in: query
        name: page
        type: integer
      - description: firstname
        in: query
        name: firstname
        type: string
      - description: lastname
        in: query
        name: lastname
        type: string
      - description: phone
        in: query
        name: phone
        type: string
      - description: search by created_at_from
        in: query
        name: created_at_from
        type: string
      - description: search by created_at_to
        in: query
        name: created_at_to
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/user_service.ListUsersResponse'
            type: array
        "400":
          description: Bad Request
          schema:
            allOf:
            - $ref: '#/definitions/v1.Response'
            - properties:
                data:
                  type: string
              type: object
        "404":
          description: Not Found
          schema:
            allOf:
            - $ref: '#/definitions/v1.Response'
            - properties:
                data:
                  type: string
              type: object
        "500":
          description: Internal Server Error
          schema:
            allOf:
            - $ref: '#/definitions/v1.Response'
            - properties:
                data:
                  type: string
              type: object
      summary: GetAll User
      tags:
      - user
  /v1/user/update/{id}:
    put:
      consumes:
      - application/json
      description: Update an existing user with the provided details
      parameters:
      - description: User ID to update
        in: path
        name: id
        required: true
        type: integer
      - description: Updated data for the user
        in: body
        name: user
        required: true
        schema:
          $ref: '#/definitions/user_service.UpdateUsersRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/user_service.UpdateUsersRequest'
        "400":
          description: Bad Request
          schema:
            allOf:
            - $ref: '#/definitions/v1.Response'
            - properties:
                data:
                  type: string
              type: object
        "404":
          description: Not Found
          schema:
            allOf:
            - $ref: '#/definitions/v1.Response'
            - properties:
                data:
                  type: string
              type: object
        "500":
          description: Internal Server Error
          schema:
            allOf:
            - $ref: '#/definitions/v1.Response'
            - properties:
                data:
                  type: string
              type: object
      summary: Update an existing user
      tags:
      - user
securityDefinitions:
  ApiKeyAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"